ダッシュボード生成を外部関数化
